chore(devtools): refactor vfs for devtools
