+++ devtools/devtools-backend/CMakeLists.txt
@@ -147,7 +147,6 @@ set(SOURCE_SET
     src/module/request/dom_push_nodes_request.cc
     src/module/request/dom_push_node_by_path_request.cc
     src/module/util/base64.cc
-    src/module/vfs/devtools_handler.cc
     src/tunnel/net_channel.cc
     src/tunnel/tcp/frame_codec.cc
     src/tunnel/tcp/tcp_channel.cc
