+++ modules/vfs/android/src/main/cpp/CMakeLists.txt
@@ -50,18 +50,12 @@ add_subdirectory(${PROJECT_ROOT_DIR}/modules/vfs/native ${CMAKE_CURRENT_BINARY_D
 target_link_libraries(${PROJECT_NAME} PUBLIC vfs_native)
 # endregion
 
-# region devtools
-if (ENABLE_INSPECTOR)
-    GlobalPackages_Add(devtools_backend)
-    target_link_libraries(${PROJECT_NAME} PRIVATE devtools_backend)
-endif ()
-# endregion
-
 # region source set
 set(SOURCE_SET src/handler/asset_handler.cc
   src/handler/file_handler.cc
   src/handler/jni_delegate_handler.cc
-  src/handler/uri.cc)
+  src/handler/uri.cc
+  src/vfs_resource_holder.cc)
 
 target_sources(${PROJECT_NAME} PRIVATE ${SOURCE_SET})
 # endregion
