+++ modules/vfs/native/include/vfs/handler/uri_handler.h
@@ -29,8 +29,6 @@
 
 namespace hippy {
 inline namespace vfs {
-constexpr char kCallFromKey[] = "__Hippy_call_from";
-constexpr char kCallFromJavaValue[] = "java";
 
 class UriHandler {
  public:
